posts-grpc/pyproject.toml
[project]
name = "posts-grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpc-interceptor>=0.15.4",
    "grpcio>=1.74.0",
    "grpcio-reflection>=1.74.0",
    "protobuf>=6.32.0",
    "psycopg[binary,pool]>=3.2.10",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.74.0",
    "py-mon>=2.1.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.10",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--cov=.",
    "--cov-report=xml",
]

